Fix bug #29 for real now! And remove typeof from SemRep.
This patch fixes bug #29 for real, without the various breakage the
previous patch caused. In addition, it allows the third example listed
in my last patch to typecheck properly.

This change makes it no longer necessary to have a representation for
'typeof' in SemRep, so this patch removes that, as well.

darcs-hash:20090714035021-fbb9b-1b09869354e00fa792b01ee063856e599a5255b7.gz