Arrays can participate in subtraction.
Previously, subtraction was allowed on pointers but not arrays. This
might be better implemented by coercing arrays to pointers beforehand,
but this works, and is easier at the moment.

Eventually, it might be best to refactor the type checker a bit to
support array->pointer coercion in the right places.

darcs-hash:20091112193809-fbb9b-3e6547f54e1714669b0501d5325569b0698dee07.gz